net/py-foolscap: Update to 21.7.0
I believe this is a relatively minor update, plus bringing python 3.x
support.  Upstream does not publish NEWS or a changelog, and the
homepage does not even mention the release.

Update prompted by tahoe-lfas needing this (py3-compattible) version.
I am not aware of anything else needing it.  If so and this version
isn't ok, we'll have to go to versioned_dependencies and import the
old one -- but this builds ok with 2.7.