py-django3: updated to 3.2.8
Django 3.2.8 fixes two bugs in 3.2.7.

Bugfixes

Fixed a bug in Django 3.2 that caused incorrect links on read-only fields in the admin.
Fixed a regression in Django 3.2 that caused incorrect selection of items across all pages when actions were placed both on the top and bottom of the admin change-list view.